pyproject.toml
[project]
name = "onyx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.0",
    "cohere>=5.18.0",
    "email-validator>=2.3.0",
    "fastapi>=0.117.1",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "langgraph>=0.6.7",
    "matplotlib>=3.10.6",
    "openai>=1.109.0",
    "pillow>=11.3.0",
    "pinecone>=7.3.0",
    "pycryptodome>=3.23.0",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.26.4",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "tqdm>=4.67.1",
]
